refactor: split request handlers by protocol domain
