+++ src/global_state/handlers/request/formatting.rs
@@ -1,0 +1,66 @@
+﻿use crate::{
+    global_state::snapshot::GlobalStateSnapshot,
+    lsp_ext::{from_proto, to_proto},
+};
+
+pub(crate) fn handle_formatting(
+    snap: GlobalStateSnapshot,
+    params: lsp_types::DocumentFormattingParams,
+) -> anyhow::Result<Option<Vec<lsp_types::TextEdit>>> {
+    let file_id = from_proto::file_id(&snap, &params.text_document.uri)?;
+    let line_info = snap.line_info(file_id)?;
+
+    let config = formatting_config(&snap, &params.options);
+    let edit = snap
+        .analysis
+        .format(file_id, None, &line_info, config, snap.cancellation.clone())?
+        .map_err(to_proto::format_error)?;
+
+    let text_edits = edit.map(|edit| to_proto::text_edits(&line_info, edit));
+    Ok(text_edits)
+}
+
+pub(crate) fn handle_range_formatting(
+    snap: GlobalStateSnapshot,
+    params: lsp_types::DocumentRangeFormattingParams,
+) -> anyhow::Result<Option<Vec<lsp_types::TextEdit>>> {
+    let file_id = from_proto::file_id(&snap, &params.text_document.uri)?;
+    let line_info = snap.line_info(file_id)?;
+    let line_ranges =
+        Some((params.range.start.line as usize)..((params.range.end.line as usize) + 1));
+
+    let config = formatting_config(&snap, &params.options);
+    let edit = snap
+        .analysis
+        .format(file_id, line_ranges, &line_info, config, snap.cancellation.clone())?
+        .map_err(to_proto::format_error)?;
+
+    let text_edits = edit.map(|edit| to_proto::text_edits(&line_info, edit));
+    Ok(text_edits)
+}
+
+pub(crate) fn handle_on_type_formatting(
+    snap: GlobalStateSnapshot,
+    params: lsp_types::DocumentOnTypeFormattingParams,
+) -> anyhow::Result<Option<Vec<lsp_types::TextEdit>>> {
+    let position = from_proto::file_position(&snap, params.text_document_position)?;
+    let line_info = snap.line_info(position.file_id)?;
+
+    let config = formatting_config(&snap, &params.options);
+    let edit = snap
+        .analysis
+        .format_on_type(position, params.ch, &line_info, config, snap.cancellation.clone())?
+        .map_err(to_proto::format_error)?;
+
+    let text_edits = edit.map(|edit| to_proto::text_edits(&line_info, edit));
+    Ok(text_edits)
+}
+
+fn formatting_config(
+    snap: &GlobalStateSnapshot,
+    options: &lsp_types::FormattingOptions,
+) -> ide::formatting::FmtConfig {
+    let mut config = snap.config.fmt();
+    config.apply_editor_options(options.tab_size, options.insert_spaces);
+    config
+}
