+++ src/global_state/handlers/request/navigation.rs
@@ -1,0 +1,112 @@
+﻿use ide::{FileRange, references::References};
+use itertools::Itertools;
+
+use crate::{
+    global_state::snapshot::GlobalStateSnapshot,
+    lsp_ext::{from_proto, to_proto},
+};
+
+pub(crate) fn handle_goto_definition(
+    snap: GlobalStateSnapshot,
+    params: lsp_types::GotoDefinitionParams,
+) -> anyhow::Result<Option<lsp_types::GotoDefinitionResponse>> {
+    let position = from_proto::file_position(&snap, params.text_document_position_params)?;
+    let Some(nav_info) = snap.analysis.goto_definition(position)? else {
+        return Ok(None);
+    };
+
+    let src = FileRange { file_id: position.file_id, range: nav_info.range };
+    let res = to_proto::goto_definition_response(&snap, Some(src), nav_info.info)?;
+    Ok(Some(res))
+}
+
+pub(crate) fn handle_goto_declaration(
+    snap: GlobalStateSnapshot,
+    params: lsp_types::request::GotoDeclarationParams,
+) -> anyhow::Result<Option<lsp_types::request::GotoDeclarationResponse>> {
+    let position = from_proto::file_position(&snap, params.text_document_position_params.clone())?;
+    let Some(nav_info) = snap.analysis.goto_declaration(position)? else {
+        return handle_goto_definition(snap, params);
+    };
+    let src = FileRange { file_id: position.file_id, range: nav_info.range };
+    let res = to_proto::goto_definition_response(&snap, Some(src), nav_info.info)?;
+    Ok(Some(res))
+}
+
+pub(crate) fn handle_document_highlight(
+    snap: GlobalStateSnapshot,
+    params: lsp_types::DocumentHighlightParams,
+) -> anyhow::Result<Option<Vec<lsp_types::DocumentHighlight>>> {
+    let position = from_proto::file_position(&snap, params.text_document_position_params)?;
+    let line_info = snap.line_info(position.file_id)?;
+    let config = snap.config.document_highlight();
+    let Some(highlights) = snap.analysis.document_highlight(position, config)? else {
+        return Ok(None);
+    };
+
+    let res = highlights
+        .into_iter()
+        .map(|highlight| to_proto::document_highlight(&line_info, highlight))
+        .collect();
+    Ok(Some(res))
+}
+
+pub(crate) fn handle_references(
+    snap: GlobalStateSnapshot,
+    params: lsp_types::ReferenceParams,
+) -> anyhow::Result<Option<Vec<lsp_types::Location>>> {
+    let include_declaration =
+        params.context.include_declaration && snap.config.references_include_declaration();
+    let position = from_proto::file_position(&snap, params.text_document_position)?;
+    let config = snap.config.references();
+    let Some(refs) = snap.analysis.references(position, config)? else {
+        return Ok(None);
+    };
+    let partial_issue_count: usize =
+        refs.iter().map(|references| references.status.issue_count()).sum();
+    if partial_issue_count > 0 {
+        tracing::debug!(partial_issue_count, "references result is partial");
+    }
+
+    let locations = refs
+        .into_iter()
+        .flat_map(|References { def, refs, .. }| {
+            let decl = if include_declaration { def.unwrap_or_default() } else { Vec::new() }
+                .into_iter()
+                .map(|nav| FileRange { file_id: nav.file_id, range: nav.focus_or_full_range() });
+
+            let refs = refs.into_iter().flat_map(|(file_id, refs)| {
+                refs.into_iter().map(move |(range, _)| FileRange { file_id, range })
+            });
+
+            decl.chain(refs)
+        })
+        .unique()
+        .filter_map(|frange| to_proto::location(&snap, frange).ok())
+        .collect_vec();
+
+    Ok(Some(locations))
+}
+
+pub(crate) fn handle_hover(
+    snap: GlobalStateSnapshot,
+    params: lsp_types::HoverParams,
+) -> anyhow::Result<Option<lsp_types::Hover>> {
+    let position = from_proto::file_position(&snap, params.text_document_position_params)?;
+
+    let config = snap.config.hover();
+    let hover_format = config.format;
+    let Some(hover_info) = snap.analysis.hover(position, config)? else {
+        return Ok(None);
+    };
+
+    let line_info = snap.line_info(position.file_id)?;
+    let range = to_proto::range(&line_info, hover_info.range);
+
+    let res = lsp_types::Hover {
+        contents: to_proto::hover_contents(hover_info.info, hover_format),
+        range: Some(range),
+    };
+
+    Ok(Some(res))
+}
